Module CharEncoding.Naive

Basic charset conversion. The conversion routine implemented in this module is not able to detect encoding. We recommend using a library such as camomile for a more complete solution.

val convert : recode

Convert charset.